requirements/bromine.txt
# Product dependencies
Pillow
PyHamcrest
requests
selenium
six
